various updates

+++ scripts/add_electricity.py
@@ -511,7 +511,7 @@ def move_generators(
         .str.replace(" offwind-\w+", "", regex=True)
     ).rename('name')
     
-    move_generators = pd.merge(move_generators, cluster_map, left_on='name', right_index=True)
+    move_generators = pd.merge(move_generators.astype(int), cluster_map, left_on='name', right_index=True)
 
     move_generators = move_generators[move_generators.isin(n.buses.index)]
     n.generators.loc[move_generators.index, "bus"] = move_generators.busmap
